Add log variables $http_all and $sent_http_all
These are meant to log the entire request and response
headers, respectively.

There are cases when we want to log the whole request (or response)
headers, for example, when we don't know in advance the
field the client sends. Currently we must know exactly the
header fields we want to log.

This patch adds these two variables that contains

$http_all: all request headers as received from client
$sent_http_all: all response headers as sent to the client

Closes #426.